+++ CHANGELOG.md
@@ -52,6 +52,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Azure IoT Central** (`azureiotcentral.py`) connector client with unit tests and samples
 - **Azure Event Grid** (`azureeventgrid.py`) connector client with unit tests and samples
 - **Universal Print** (`universalprint.py`) connector client with unit tests and samples
+- **Twitter** (`twitter.py`) connector client with unit tests and samples
+- **WordPress** (`wordpress.py`) connector client with unit tests and samples
+- **Plivo** (`plivo.py`) connector client with unit tests and samples
+- **Rev.ai** (`revai.py`) connector client with unit tests and samples
+- **Starmind** (`starmind.py`) connector client with unit tests and samples
+- **Tallyfy** (`tallyfy.py`) connector client with unit tests and samples
 - **TextRequest** (`textrequest.py`) connector client with unit tests and samples
 - **Trello** (`trello.py`) connector client with unit tests and samples
 - **ClickSend SMS** (`clicksendsms.py`) connector client with unit tests and samples